mafia_backend/app/requirements.txt
SQLAlchemy==2.0.36
alembic==1.14.0
pymysql==1.1.1
python-dotenv==1.0.1
mysql-connector-python==9.1.0
pandas==2.2.3
Flask==3.1.0
Flask-Cors==5.0.0
dependency-injector==4.43.0
# Flask-JWT==0.3.2
flask-talisman==1.1.0
pendulum==3.0.0
Flask-JWT-Extended==4.7.1
bcrypt==4.2.1